feat: implement LaTeX formatting utilities and refactor processing logic
